reactored Todo.js, TodoList.js
